+++ authority-portal-backend/authority-portal-db/src/main/resources/db/migration/V14__Change_Primary_Key_For_Component.sql
@@ -1,0 +1,4 @@
+alter table component drop constraint component_pkey;
+update component set name = id;
+update component set id = client_id;
+alter table component add primary key (id, environment);